jemalloc: workaround exceptions
If a sealed capability is passed, performing masking as part of
looking up the object may cause an exception on version 8
of the CHERI-RISC-V ISA.  Since this code only cares about addresses
just create a NULL-derived uintptr_t and avoid this issue.  The sealed
capability will be rejected later in unbound_ptr.